add StateChecker
Added StateChecker class to the front end
