--- v2/main_app/fixed_arena_app/static/js/inject_vuln.js
@@ -1,1 +1,0 @@
-//
