+++ v2/main_app/fixed_arena_app/templates/controller_macros.html
@@ -136,6 +136,4 @@
     {% endfor %}
     </tbody>
 {% endmacro %}
-
-
 <!-- [ eof ] -->
