+++ v2/main_app/fixed_arena_app/templates/fixed_arena_class.html
@@ -1,6 +1,6 @@
 {% extends "fa_base.html" %}
 {% import 'cyber_learn_logo.html' as logo %}
-{% import "vuln_macros.html" as vuln_macros %}
+{% import "controller_macros.html" as ctrl_macros %}
 {% import "fixed_arena_macros.html" as fixed_arena_macros %}
 {% block title %}{{ fixed_arena_class.summary.name }}{% endblock title %}
 {% block nav_title %}{{ fixed_arena_class.summary.name }}{% endblock nav_title %}
@@ -30,7 +30,9 @@
 {% endblock %}
 {% block content %}
     <script type="text/javascript" src='{{ url_for("static", filename="js/login.js") }}'></script>
-    <script type="text/javascript" src="{{ url_for('fixed_arena.static', filename='js/fixed_arena_utils.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('fixed_arena.static', filename='js/fixed_arena.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('fixed_arena.static', filename='js/state_checker.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('fixed_arena.static', filename='js/controller.js') }}"></script>
     <script>
         var teacher_email = null;
         var auth_config = {
@@ -120,6 +122,7 @@
                         </div>
                     </div>
                     <div id="workspace-servers" class="info-container">
+                        <div class="loadAnimationWrapper"><div id="loadAnimation"></div></div>
                         <div class="server-manage-wrapper">
                             <div class="server-toolbar-div">
                                 <ul class="server-toolbar">
@@ -152,14 +155,17 @@
                                     <!--<td id="state-container-{{workspace.key.name}}" class="status-indicator-container" data-toggle="tooltip" title="{{workspace.state}}">
                                         <span class="status-indicator" id="status-indicator-{{workspace.key.name}}" data-workout-state="{{workspace.state}}">
                                         </span><i class="fa fa-stop-circle" style="color: #1d8037; font-size: 36px;"></i></td>-->
-                                    {% if fixed_arena_class.state == 50 %}
-                                    <!-- if fixed_arena_class.state == 4 -->
-                                        <td><i class="fa fa-circle" style="color: #1d8037; font-size: 20px;"></i></td>
-                                    {% elif fixed_arena_class.state == 53 %} <!-- workspace stopped == 6 -->
-                                        <td><i class="fa fa-stop-circle" style="color: #838683; font-size: 20px;"></i></td>
-                                    {% else %}
-                                        <td><i class="fa fa-circle" style="color: #b3be27; font-size: 20px;"></i></td>
-                                    {% endif %}
+                                    <td>
+                                        {% if fixed_arena_class.state == 50 %}
+                                            <i class="fa fa-circle buildState running" data-toggle="tooltip" data-placement="bottom" title="RUNNING"></i>
+                                        {% elif fixed_arena_class.state == 53 %}
+                                            <i class="fa fa-stop-circle buildState stopped" data-toggle="tooltip" data-placement="bottom" title="STOPPED"></i>
+                                        {% elif fixed_arena_class.state == 70 %}
+                                            <i class="fa fa-stop-circle buildState deleted" data-toggle="tooltip" data-placement="bottom" title="DELETED"></i>
+                                        {% else %}
+                                            <i class="fa fa-circle buildState transition" data-toggle="tooltip" data-placement="bottom" title="WORKING"></i>
+                                        {% endif %}
+                                    </td>
                                     <td><a class='workspace-link' href="{{ url_for('student_app.fixed_arena_student', build_id=workspace.key.name) }}">{{ workspace.key.name }}</a></td>
                                     <td id="workspace_student_name_{{ student_count }}">{{ workspace.student_name }}</td>
                                     {% if fixed_arena_class.workspace_settings.registration_required %}
@@ -207,7 +213,7 @@
                         <!--Holds both form builder and attack results table-->
                         <div id="vuln-templates-container">
                              <table class="table table-hover" id="vuln-templates-table">
-                                 {{ vuln_macros.display_vulns(attack_spec.attack) }}
+                                 {{ ctrl_macros.display_vulns(attack_spec.attack) }}
                              </table>
                             <div id="vuln-template-btn-div">
                                 <button id='vuln-template-btn' class="btn btn-sm" type="button" disabled hidden>Select Template</button>
@@ -225,7 +231,7 @@
                                     <div class="modal-body">
                                         <form id="vuln-template-form" name="vuln-template-form" method="post" action="/api/inject/">
                                             <div id="vuln-template-args"></div>
-                                            {{ vuln_macros.target_network('vuln-template-form', workspaces, fixed_arena_class.workspace_settings.registration_required) }}
+                                            {{ ctrl_macros.target_network('vuln-template-form', workspaces, fixed_arena_class.workspace_settings.registration_required) }}
                                         </form>
                                     </div>
                                     <div class="modal-footer">
